requirements.txt
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/pytest-sina-framework.git@master#egg=pytest-sina-framework
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/pytest-rst-report.git@master#egg=pytest_rst_report
pytest_check == 2.1.4
webdriver_manager == 4.0.1
selenium == 4.20.0



